requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
pyTelegramBotAPI==4.7.0
requests==2.28.1
urllib3==1.26.12

python-dotenv~=0.21.0